Files Commit
A skeleton POC RAML with a little bit of meat. I started with RAML 0.8
but soon found out that RAML 1.0 is sooo much better. Reminder to self
keep up to date with this specification.

In this commit we are essentially putting the POC in place hoewever
there are some rough edges such as filtering traits that needs to be
looked at.

ALso I have not tested this with RAML to code generation so be gentle if
it does not work.